requirements.txt
flask==3.0.0
flask-cors==4.0.0
boto3==1.34.0
pymongo==4.6.0
python-dotenv==1.0.0
tensorflow==2.15.0
ultralytics==8.1.0
pillow==10.2.0
shapely==2.0.2
numpy==1.26.3